fixed bug of detail product

+++ app/src/main/java/com/example/myshop/model/ProductDataClass.kt
@@ -8,7 +8,7 @@ data class ProductDataClass(
     var id: String = "",
     var FirstName: String = "",
     var product_title: String = "",
-    var product_pricce: Float = 0f,
+    var product_pricce: Int = 0,
     var product_desc: String = "",
     var product_quantity: Int = 0,
     var product_image: String = "",
